+++ src/main/java/pl/pw/radeja/speex/encoders/HideF0EncoderFirstLastRand.java
@@ -5,7 +5,7 @@ import pl.pw.radeja.speex.pitch.changers.LinearApproximateRand;
 
 public class HideF0EncoderFirstLastRand extends HideF0EncoderFirstLast {
 
-    public HideF0EncoderFirstLastRand(int threshold, String path) {
+    public HideF0EncoderFirstLastRand(float threshold, String path) {
         super(new LinearApproximateRand(threshold, 1L), path);
         this.type = Config.HideF0Type.FIRST_LAST_RAND;
     }
