+++ src/main/java/pl/pw/radeja/speex/pitch/collectors/CalculatedThresholdPrinter.java
@@ -6,11 +6,11 @@ import java.util.stream.Collectors;
 public class CalculatedThresholdPrinter {
 
     public static void print(List<PitchCollector> pitchCollectors) {
-        Map<String, Map<Integer, Integer>> sampleNameToThresholdToPithcCollectors = getResultSet(pitchCollectors);
-        Map<String, Map<Integer, Integer>> getResultSetAfterHideF0 = getResultSetAfterHideF0(pitchCollectors);
+        Map<String, Map<Float, Integer>> sampleNameToThresholdToPithcCollectors = getResultSet(pitchCollectors);
+        Map<String, Map<Float, Integer>> getResultSetAfterHideF0 = getResultSetAfterHideF0(pitchCollectors);
         ArrayList<String> sampleNames = new ArrayList<>(sampleNameToThresholdToPithcCollectors.keySet());
         Collections.sort(sampleNames);
-        ArrayList<Integer> thresholds = new ArrayList<>(sampleNameToThresholdToPithcCollectors.get(sampleNames.get(0)).keySet());
+        ArrayList<Float> thresholds = new ArrayList<>(sampleNameToThresholdToPithcCollectors.get(sampleNames.get(0)).keySet());
         Collections.sort(thresholds);
         System.out.println("\n\n\n");
         System.out.println("Calculated Thresholds");
@@ -34,9 +34,9 @@ public class CalculatedThresholdPrinter {
         System.out.println("\n\n\n");
     }
 
-    private static Map<String, Map<Integer, Integer>> getResultSet(List<PitchCollector> result) {
-        List<Integer> thresholds = result.stream().map(PitchCollector::getThreshold).sorted().collect(Collectors.toList());
-        Map<String, Map<Integer, Integer>> resultMap = new HashMap<>();
+    private static Map<String, Map<Float, Integer>> getResultSet(List<PitchCollector> result) {
+        List<Float> thresholds = result.stream().map(PitchCollector::getThreshold).sorted().collect(Collectors.toList());
+        Map<String, Map<Float, Integer>> resultMap = new HashMap<>();
         result.forEach(r -> {
             resultMap.putIfAbsent(r.getSampleName(), new HashMap<>());
             resultMap.get(r.getSampleName()).putIfAbsent(r.getThreshold(), countThresholdFromLinear(r, thresholds));
@@ -44,9 +44,9 @@ public class CalculatedThresholdPrinter {
         return resultMap;
     }
 
-    private static Map<String, Map<Integer, Integer>> getResultSetAfterHideF0(List<PitchCollector> result) {
-        List<Integer> thresholds = result.stream().map(PitchCollector::getThreshold).sorted().collect(Collectors.toList());
-        Map<String, Map<Integer, Integer>> resultMap = new HashMap<>();
+    private static Map<String, Map<Float, Integer>> getResultSetAfterHideF0(List<PitchCollector> result) {
+        List<Float> thresholds = result.stream().map(PitchCollector::getThreshold).sorted().collect(Collectors.toList());
+        Map<String, Map<Float, Integer>> resultMap = new HashMap<>();
         result.forEach(r -> {
             resultMap.putIfAbsent(r.getSampleName(), new HashMap<>());
             resultMap.get(r.getSampleName()).putIfAbsent(r.getThreshold(), countThresholdAfterHideF0FromLinear(r, thresholds));
@@ -54,8 +54,8 @@ public class CalculatedThresholdPrinter {
         return resultMap;
     }
 
-    private static Integer countThresholdFromLinear(PitchCollector collector, List<Integer> thresholds) {
-        Integer min = thresholds.indexOf(collector.getThreshold()) > 0 ? thresholds.get(thresholds.indexOf(collector.getThreshold()) - 1) : 0;
+    private static Integer countThresholdFromLinear(PitchCollector collector, List<Float> thresholds) {
+        Float min = thresholds.indexOf(collector.getThreshold()) > 0 ? thresholds.get(thresholds.indexOf(collector.getThreshold()) - 1) : 0f;
         return (int) collector.getFramePitchValues().stream().filter(p -> {
             if (collector.getThreshold() != 0) {
                 return p.getCalculatedThreshold() <= collector.getThreshold() && p.getCalculatedThreshold() > min;
@@ -65,8 +65,8 @@ public class CalculatedThresholdPrinter {
         }).count();
     }
 
-    private static Integer countThresholdAfterHideF0FromLinear(PitchCollector collector, List<Integer> thresholds) {
-        Integer min = thresholds.indexOf(collector.getThreshold()) > 0 ? thresholds.get(thresholds.indexOf(collector.getThreshold()) - 1) : 0;
+    private static Integer countThresholdAfterHideF0FromLinear(PitchCollector collector, List<Float> thresholds) {
+        Float min = thresholds.indexOf(collector.getThreshold()) > 0 ? thresholds.get(thresholds.indexOf(collector.getThreshold()) - 1) : 0f;
         return (int) collector.getFramePitchValues().stream().filter(p -> {
             if (collector.getThreshold() != 0) {
                 if (p.isChanged()) {
